+++ src/sgem.hpp
@@ -407,6 +407,9 @@ public:
     SparseGmpEigenMatrix rowProd() const;
     SparseGmpEigenMatrix& rowProd_new() const;
 
+    // sorting functions
+    SparseGmpEigenMatrix sort(const int& dim, const int& type, std::vector < std::vector < IndexType > >& index, std::vector < IndexType >& nbNegatives) const;
+    SparseGmpEigenMatrix& sort_new(const int& dim, const int& type, std::vector < std::vector < IndexType > >& index, std::vector < IndexType >& nbNegatives) const;